fix: refactoring code

+++ controllers/movies.js
@@ -4,11 +4,12 @@ const { HTTP_STATUS_OK } = require('../consts/consts');
 const {
   BadRequestError,
   NotFoundError,
-  // ForbiddenError,
+  ForbiddenError,
 } = require('../errors/errors');
 
 const getMovies = (req, res, next) => {
-  Movies.find({})
+  const owner = req.user._id;
+  Movies.findById(owner)
     .then((movies) => res.send(movies))
     .catch((err) => next(err));
 };
@@ -62,11 +63,11 @@ const deleteMovies = (req, res, next) => {
       if (!movies) {
         return next(new NotFoundError('Фильм по указанному _id не найден'));
       }
-      // if (movies.owner.toString() !== req.user._id) {
-      //   return next(
-      //     new ForbiddenError('Разрешено удалять только свои сохраненные фильмы'),
-      //   );
-      // }
+      if (movies.owner.toString() !== req.user._id) {
+        return next(
+          new ForbiddenError('Разрешено удалять только свои сохраненные фильмы'),
+        );
+      }
       return Movies.findByIdAndDelete(_id).then((deletedMovies) => res.status(HTTP_STATUS_OK).send(deletedMovies));
     })
     .catch((err) => {
